requirements.txt
openai>=1.12.0
langgraph>=0.0.40
langchain>=0.1.0
langchain-openai>=0.0.5
pandas>=2.0.0
python-dotenv>=1.0.0
streamlit>=1.28.0
moviepy>=1.0.3
pillow>=10.0.0
numpy>=1.24.0

